Remove the channel seed configuration and respect channel rules on each ProjectRepository.cs
Sadly this may slow things down, but this is more accurate as it follows the channel rules
defined for each project rather than whichever it finds first from the seeds.

Also improved the project loading progress bar, which was based on all projects. It now
prefilters the project list so the bar is accurate based on filtering where specified.